Core rebuild, other fixes

+++ modules/database/include/modules/database/dataman_frames.h
@@ -22,7 +22,7 @@ typedef struct dataman_sensor_frame
     vec3_t gyro1;
     vec3_t gyro2;
     vec3_t mag1;
-    float press;
+    int press;
     float temp;
     double lat;
     double lon;
